Add task grouping tests for plain console
