+++ subprojects/logging/src/integTest/groovy/org/gradle/internal/logging/console/taskgrouping/AbstractConsoleBuildSrcGroupedTaskFunctionalTest.groovy
@@ -16,7 +16,7 @@
 
 package org.gradle.internal.logging.console.taskgrouping
 
-class ConsoleBuildSrcGroupedTaskFunctionalTest extends AbstractConsoleGroupedTaskFunctionalTest {
+abstract class AbstractConsoleBuildSrcGroupedTaskFunctionalTest extends AbstractConsoleGroupedTaskFunctionalTest {
 
     private static final String HELLO_WORLD_MESSAGE = 'Hello world'
     private static final String BYE_WORLD_MESSAGE = 'Bye world'
