+++ subprojects/logging/src/integTest/groovy/org/gradle/internal/logging/console/taskgrouping/AbstractConsoleVerboseRenderingFunctionalTest.groovy
@@ -1,0 +1,57 @@
+/*
+ * Copyright 2018 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.internal.logging.console.taskgrouping
+
+abstract class AbstractConsoleVerboseRenderingFunctionalTest extends AbstractConsoleVerboseBasicFunctionalTest {
+    def 'failed task result can be rendered (#consoleType console)'() {
+        given:
+        buildFile << '''
+task myFailure {
+    doLast {
+        assert false
+    }
+}
+'''
+        when:
+        executer.withConsole(consoleType)
+        fails('myFailure')
+
+        then:
+        result.groupedOutput.task(':myFailure').outcome == 'FAILED'
+    }
+
+    def 'up-to-date task result can be rendered'() {
+        given:
+        buildFile << '''
+task upToDate{
+    outputs.upToDateWhen {true}
+    doLast {}
+}
+'''
+        when:
+        succeeds('upToDate')
+
+        then:
+        result.groupedOutput.task(':upToDate').outcome == null
+
+        when:
+        succeeds('upToDate')
+
+        then:
+        result.groupedOutput.task(':upToDate').outcome == 'UP-TO-DATE'
+    }
+}
